Update scripts, tmux, zsh
- alacritty: columns from 130 to 160
- bottom: mind keybinds dawg
- tmux: update config, add few keybinds
- .zshrc: comment out Termux theme
- scripts:
  - RasPiOS: v0.1.2, lil tweaks, cargo-update
  - recon_nmap: v0.1.7, vars/funcs rework for multidistro + good IP, logging...
- zsh: styling

+++ scripts/setup/cheat-symlink.sh
@@ -1,24 +1,33 @@
 #!/bin/bash
 
-dotpath="$HOME/dotfiles/.config/cheat"
 cheatpath="$HOME/.config/cheat"
+dotpath="$HOME/dotfiles/.config/cheat"
 
-if command -v eza &>/dev/null
-    then print="eza --icons -laI .git --no-user --no-permissions --no-filesize --git -TL 2"
-  elif command -v tree &>/dev/null
-    then print="tree -L 2"
-  else print="ls -l"
+if command -v eza &>/dev/null; then
+    print="eza --icons -laI .git --no-user --no-permissions --no-filesize --git -TL 2"
+elif command -v exa &>/dev/null; then
+    print="exa --icons -laI .git --no-user --no-permissions --no-filesize --git -TL 2"
+elif command -v tree &>/dev/null; then
+    print="tree -L 2"
+else
+    print="ls -l"
 fi
 
-echo -e "This is '~/.config/cheat' BEFORE the script:"
-$print $cheatpath
 
-rm $cheatpath/conf.yml
-ln -s $dotpath/conf.yml $cheatpath/
+if [ -L $cheatpath/conf.yml ]; then
+    echo "It seems there's already some symlinks at '~/.config/cheat'. Exiting..."
+    exit 1
+else
+    echo -e "This is '~/.config/cheat' BEFORE the script:"
+    $print $cheatpath
 
-rmdir $cheatpath/cheatsheets/personal
-ln -s $dotpath/cheatsheets/personal $cheatpath/cheatsheets/
-ln -s $dotpath/cheatsheets/wip $cheatpath/cheatsheets/
+    rm $cheatpath/conf.yml &&
+        ln -s $dotpath/conf.yml $cheatpath/
 
-echo -e "\nThis is '~/.config/cheat' AFTER the script:"
-$print $cheatpath
+    rmdir $cheatpath/cheatsheets/personal &&
+        ln -s $dotpath/cheatsheets/wip $cheatpath/cheatsheets/ &&
+        ln -s $dotpath/cheatsheets/personal $cheatpath/cheatsheets/
+
+    echo -e "\nThis is '~/.config/cheat' AFTER the script:"
+    $print $cheatpath
+fi
